requirements.txt
fastapi
uvicorn
sentence-transformers
scikit-learn
numpy
pydantic
python-multipart
pandas
nltk
PyPDF2
python-docx